@@ -1,6 +1,5 @@
 library new_post_view;
 
-import 'dart:async';
 import 'dart:io';
 
 import 'package:flutter/material.dart';
@@ -10,14 +9,11 @@ import 'package:responsive_builder/responsive_builder.dart';
 import 'package:stacked/stacked.dart';
 import 'package:tribes/core/tribe/new_post/new_post_view_model.dart';
 import 'package:tribes/models/tribe_model.dart';
-import 'package:tribes/services/firebase/database_service.dart';
-import 'package:tribes/services/firebase/storage_service.dart';
 import 'package:tribes/shared/constants.dart' as Constants;
 import 'package:tribes/shared/decorations.dart' as Decorations;
 import 'package:tribes/shared/widgets/custom_awesome_icon.dart';
 import 'package:tribes/shared/widgets/custom_button.dart';
 import 'package:tribes/shared/widgets/custom_scroll_behavior.dart';
-import 'package:tribes/shared/widgets/discard_changes_dialog.dart';
 import 'package:tribes/shared/widgets/loading.dart';
 
 part 'new_post_view_[mobile].dart';
@@ -36,6 +32,7 @@ class _NewPostViewState extends State<NewPostView> {
     return ViewModelBuilder<NewPostViewModel>.reactive(
       viewModelBuilder: () => NewPostViewModel(),
       onModelReady: (model) => model.initState(
+        context: context,
         tribe: widget.tribe,
         isMounted: this.mounted,
       ),
