@@ -13,10 +13,10 @@ import 'package:tribes/core/tribe/post_room/post_room_view_model.dart';
 import 'package:tribes/core/tribe/widgets/fullscreen_carousel.dart';
 import 'package:tribes/models/post_model.dart';
 import 'package:tribes/models/user_model.dart';
-import 'package:tribes/services/firebase/database_service.dart';
 import 'package:tribes/shared/constants.dart' as Constants;
 import 'package:tribes/shared/widgets/custom_awesome_icon.dart';
 import 'package:tribes/shared/widgets/custom_scroll_behavior.dart';
+import 'package:tribes/shared/widgets/discard_changes_dialog.dart';
 import 'package:tribes/shared/widgets/like_button.dart';
 import 'package:tribes/shared/widgets/posted_date_time.dart';
 import 'package:tribes/shared/widgets/user_avatar.dart';
